Update python tutorials
