+++ python/exemples/gt_non_recours/framework.py
@@ -8,7 +8,6 @@ from reformes.reforme_test import MaReform
 legislation_france = openfisca_france.FranceTaxBenefitSystem()
 
 reformes = {
-    "France" : legislation_france,  # situation actuelle
     "Reforme de test" : MaReform(legislation_france)
 }
 
@@ -21,10 +20,9 @@ situations = "situations/"
 period_month = '2018-02'
 period_year = '2017'
 calculs = {
-    # 'base_ressource_mensuelle_individu': period_month,
-    # 'base_ressource_mensuelle_famille': period_month,
-    # 'base_ressource_annuelle_individu': period_year,
-    # 'base_ressources_al_2019': period_month,
+    'base_ressource_mensuelle_individu': period_month,
+    'base_ressource_mensuelle_famille': period_month,
+    'base_ressource_annuelle_individu': period_year,
     'aide_logement_base_ressources': period_month,
     'aide_logement': period_month,
     'loyer': period_month
