+++ python/reforme_parametre.py
@@ -1,9 +1,6 @@
 # -*- coding: utf-8 -*-
-
-import openfisca_france
-from openfisca_core.simulations import Simulation
-from openfisca_core import reforms
-from openfisca_core import periods
+from openfisca_core import reforms, periods
+from openfisca_france.model.base import *
 
 ####### Décrivez votre réforme ###########
 
@@ -20,19 +17,3 @@ def modifier_un_parametre(parameters):
 class MaReform(reforms.Reform):
     def apply(self):
         self.modify_parameters(modifier_function = modifier_un_parametre)
-
-####### Consulter la situation actuelle ##############
-legislation_france = openfisca_france.FranceTaxBenefitSystem()
-
-resultat_actuel = legislation_france.parameters.impot_revenu.bareme[1].rate
-
-print "Résultat actuel"
-print resultat_actuel
-
-####### Consulter la situation avec la reforme ##############
-legislation_reforme = MaReform(legislation_france)
-
-resultat_apres_reforme = legislation_reforme.parameters.impot_revenu.bareme[1].rate
-
-print "Resultat après reforme"
-print resultat_apres_reforme
